+++ client/src/components/reservation/Reservations.tsx
@@ -1,18 +1,31 @@
 import React, { useCallback, useEffect, useState } from "react";
 import { connect } from "react-redux";
 import "./Reservations.css";
-import { Button } from "@material-ui/core";
 import { DataTable } from "../tableForData/DataTable";
 import {
-  acceptReservationRequest,
-  deleteReservationRequest,
-  getAllReservationRequest,
+    acceptReservationRequest,
+    deleteReservationRequest,
+    getAllReservationRequest, getAllReservations,
 } from "../../api/reservations";
 import { RowData, Scheme } from "../tableForData/types";
 import { getScheme } from "./reservationsConfig";
-import {fetchDataTimer} from "../../App";
+import {setUpAnIntervalPollingOfFunction} from "../utils";
 
-const Reservations = (): JSX.Element => {
+
+const saveNewData = (el: RowData) => {
+    return {
+        ...el,
+        name: el['personName'],
+        phone: el['personPhone'],
+        tableCandidateId: el['assignedTableId'],
+        time: {
+            from: el.from,
+            to: el.to,
+        },
+    };
+};
+
+const Reservations = (props: {forAccept?: boolean}): JSX.Element => {
   const [scheme, setScheme] = useState<Scheme | null>(null);
   const [data, setData] = useState<RowData[] | null>(null);
 
@@ -22,49 +35,46 @@ const Reservations = (): JSX.Element => {
     );
   }, []);
   const onDelete = useCallback((event) => {
-    deleteReservationRequest(event.currentTarget.value).then(() =>
+      if(window.confirm("Вы уверены, что хотите удалить запрос на бронирование?"))
+          deleteReservationRequest(event.currentTarget.value).then(() =>
       setData(null)
     );
   }, []);
 
   const updateData = useCallback(() => {
-        getAllReservationRequest().then(
-            (receivedData) =>
-                receivedData &&
-                setData(
-                    receivedData.requests.map((el) => {
-                        return {
-                            ...el,
-                            time: {
-                                from: el.from,
-                                to: el.to,
-                            },
-                        };
-                    })
-                )
-        );
-    }, []);
+      if(props.forAccept) {
+          getAllReservationRequest().then(
+              (receivedData) =>
+                  receivedData && setData(receivedData.requests
+                      .map((el: RowData) => {
+                          return {
+                              ...el,
+                              time: {
+                                  from: el.from,
+                                  to: el.to,
+                              },
+                          };
+                  }))
+          );
+      } else {
+          getAllReservations().then(
+              (receivedData) =>
+                  receivedData && setData(receivedData.reservations.map(saveNewData))
+          );
+      }
+    }, [props.forAccept]);
 
   useEffect(() => {
-    setScheme(getScheme({ onDelete, onAccept }));
-  }, [onDelete, onAccept]);
+    setScheme(getScheme({ onDelete, onAccept, forAccept: props.forAccept }));
+  }, [onDelete, onAccept, props.forAccept]);
 
-  useEffect(() => {
-    updateData();
-    const interval = setInterval(updateData, fetchDataTimer);
-    return () => clearInterval(interval);
-  }, []);
+  useEffect(() => setUpAnIntervalPollingOfFunction(updateData), [props.forAccept]);
 
   return (
     <div className="reservationsWrapper">
-      <div className="header panelTitle">Бронирования</div>
+      <div className="header panelTitle">Запросы бронирования</div>
       <div className="tableWrapper">
         {scheme && <DataTable scheme={scheme} data={data} />}
-      </div>
-      <div className="btnWrapper">
-        <Button variant="contained" color="primary" onClick={() => {}}>
-          ОК
-        </Button>
       </div>
     </div>
   );
